ref(filestructure): rename auth to Auth
