+++ discordwrap/__init__.py
@@ -1,5 +1,5 @@
 from . import core
-from .auth import *
+from .Auth import *
 from .methods import *
 
 __version__ = "0.4.0"
